requirements.txt
bleach~=5.0.1
coverage~=7.0.2
flask-security-too[common]~=5.0.2
Flask-SQLAlchemy~=3.0.2
Flask==2.2.2
flask_admin~=1.6.0
flake8~=6.0.0
pytest-cov~=4.0.0
pytest-timeout~=2.1.0
pytest~=7.2.0
python-decouple~=3.6
werkzeug~=2.2.2
